test focal loss

+++ FocalLoss.py
@@ -8,7 +8,7 @@ class FocalLoss(nn.Module):
         super(FocalLoss, self).__init__()
         self.gamma = gamma
         self.alpha = alpha
-        if isinstance(alpha,(float,int, long)): self.alpha = torch.Tensor([alpha,1-alpha])
+        if isinstance(alpha,(float,int, int)): self.alpha = torch.Tensor([alpha,1-alpha])
         if isinstance(alpha,list): self.alpha = torch.Tensor(alpha)
         self.size_average = size_average
